Move featues directory; refactor behave environment handling
